+++ feature/createtodolist/src/main/java/com/godlife/createtodolist/CreateTodoListScreen2.kt
@@ -31,7 +31,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.godlife.designsystem.CleanArchitectureTheme
+import com.godlife.designsystem.GodLifeTheme
 import com.godlife.designsystem.GreyWhite
 import com.godlife.designsystem.PurpleMain
 
@@ -40,7 +40,7 @@ import com.godlife.designsystem.PurpleMain
 fun CreateTodoListScreen2(
     createViewModel: CreateViewModel = hiltViewModel()
 ){
-    CleanArchitectureTheme{
+    GodLifeTheme{
 
         Column(
             modifier = Modifier
@@ -127,7 +127,7 @@ fun CreateTodoListScreen2(
 @Preview(showBackground = true)
 @Composable
 fun CreateTodoListScreen2Preview(){
-    CleanArchitectureTheme{
+    GodLifeTheme{
 
         Column(
             modifier = Modifier
