+++ feature/login/src/main/java/com/godlife/login/SignUpScreen.kt
@@ -1,0 +1,20 @@
+package com.godlife.login
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import com.godlife.designsystem.GodLifeTheme
+
+@Preview(showBackground = true)
+@Composable
+fun SignUpScreen(){
+    GodLifeTheme {
+        Column(
+            modifier = Modifier.fillMaxSize()
+        ) {
+
+        }
+    }
+}
